refactor(ui/web): use presenter container convention on navbar
